float으로 수정

+++ MySQL/full_schema.sql
@@ -105,7 +105,7 @@ CREATE TABLE historical_asset_data (
 CREATE TABLE api (
     api_num INT PRIMARY KEY AUTO_INCREMENT,
     api_name VARCHAR(50),
-    api_value INT,
+    api_value DECIMAL(10,2),
     api_date DATE,
     api_datetime DATETIME,
     UNIQUE KEY uq_api_name_date (api_name, api_date)
